Move browser polyfills to platform/browser
